+++ webpack.config.js
@@ -72,8 +72,6 @@ module.exports = [
       }
       */
       
-      
-      
 
    }
 ];